% Generated by roxygen2 (4.0.2): do not edit by hand
\name{align.folder}
\alias{align.folder}
\title{Align Sequences in a Folder}
\usage{
align.folder(input.folder, output.folder = file.path(input.folder, "Aligned"),
  align.func, ...)
}
\arguments{
\item{input.folder}{a folder containing sequences.}

\item{output.folder}{older where alignments will go.}

\item{align.func}{an alignment function like \code{\link{align.clustal}} or \code{\link{align.mafft}}.}

\item{...}{parameters to \code{align.func}.}
}
\description{
Aligns all sequences ending in ".fa*" within a folder
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

