% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allele.split}
\alias{allele.split}
\title{Split Alleles For Diploid Data}
\usage{
allele.split(x)
}
\arguments{
\item{x}{a matrix or data.frame containing diploid data. Every column represents one locus with two alleles.}
}
\value{
matrix with alleles for each locus in one column split into separate columns.
}
\description{
Split loci from one column to two columns in a matrix of diploid data.
}
\note{
Alleles should be of equal length (e.g., 145095 = 145 and 095, or AG is A and G).
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

