% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allelic.richness}
\alias{allelic.richness}
\title{Allelic Richness}
\usage{
allelic.richness(g, use.na.rows = FALSE)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{use.na.rows}{logical. FALSE deletes any row with an NA for either allele.
  TRUE uses all unique (non-NA) alleles in frequency.}
}
\description{
Calculate allelic richness for diploid data.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

