% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicRichness.R
\name{allelicRichness}
\alias{allelicRichness}
\title{Allelic Richness}
\usage{
allelicRichness(g, by.strata = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{logical - return results grouped by strata?}
}
\value{
a data frame with the allelic richness of each locus calculated
  as the number of alleles divided by the number of samples without
  missing data at that locus.
}
\description{
Calculate allelic richness for each locus.
}
\examples{
data(msats.g)
allelicRichness(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
