% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.gtypes.R
\docType{methods}
\name{as.matrix,gtypes-method}
\alias{as.matrix,gtypes-method}
\alias{as.matrix.gtypes}
\title{Convert \code{gtypes} To \code{matrix}}
\usage{
\S4method{as.matrix}{gtypes}(x, one.col = FALSE, sep = "/", ids = TRUE,
  strata = TRUE, ...)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}

\item{one.col}{logical. If \code{TRUE}, then result has one column per 
locus.}

\item{sep}{character to use to separate alleles if \code{one.col} is 
\code{TRUE}.}

\item{ids}{logical. include a column for individual identifiers (\code{ids})?}

\item{strata}{logical. include a column for current statification (\code{strata})?}

\item{...}{additional arguments ot be passed to or from methods.}
}
\value{
A \code{matrix} with one row per sample.
}
\description{
Create a matrix from a \linkS4class{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noa.gov}
}
\seealso{
\link{df2gtypes}
}

