% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.multidna.R
\name{as.multidna}
\alias{as.multidna}
\title{Convert to multidna}
\usage{
as.multidna(x, ...)
}
\arguments{
\item{x}{a valid set of sequences: character matrix, list of 
character vectors, \code{DNAbin} object or list of them,
\linkS4class{gtypes} object, or \code{multidna} object.}

\item{...}{arguments to pass to \code{\link{getSequences}} if \code{x} is a
\linkS4class{gtypes} object.}
}
\description{
Convert a set of sequences to a multidna object if possible.
}
\examples{
# convert list of character vectors
data(dolph.seqs)
list.mdna <- as.multidna(dolph.seqs)
list.mdna

# convert gtypes object
data(dloop.g)
gtype.mdna <- as.multidna(dloop.g)
gtype.mdna

}
\seealso{
\link{getSequences}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
