% Generated by roxygen2 (4.0.2): do not edit by hand
\name{base.freqs}
\alias{base.freqs}
\title{Base Frequencies}
\usage{
base.freqs(x, bases = NULL)
}
\arguments{
\item{x}{a \code{\link{gtypes}} object with aligned sequences or a list of aligned DNA sequences.}

\item{bases}{character vector of bases. Must contain valid IUPAC codes. If NULL, will return summary
  of frequencies of observed bases.}
}
\value{
list containing:
\tabular{ll}{
  \code{site.freqs} \tab a matrix of base frequencies at each site.\cr
  \code{base.freqs} \tab a vector of overall base frequencies.\cr
}
}
\description{
Calculate nucleotide base frequencies along a sequence.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

