% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bowhead.snp.position}
\alias{bowhead.snp.position}
\title{Bowhead Whale SNP Genotype Groups}
\format{A data.frame of position information for SNPs to be phased}
\usage{
data(bowhead.snp.position)
}
\description{
Bowhead Whale SNP Genotype Groups
}
\references{
Morin et al.
}
\keyword{datasets}

