% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bowhead.snps}
\alias{bowhead.snps}
\title{Bowhead Whale SNP Genotypes}
\format{A data.frame of 42 SNPs with sample ids and stratification}
\usage{
data(bowhead.snps)
}
\description{
Bowhead Whale SNP Genotypes
}
\references{
Morin et al.
}
\keyword{datasets}

