% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decode.R, R/decode.loci.R, R/decode.sequences.R, R/decode.strata.R
\name{decode}
\alias{decode}
\alias{decode.loci}
\alias{decode.sequences}
\alias{decode.strata}
\title{Decode \code{gtypes}}
\usage{
decode(g)

decode.loci(g)

decode.sequences(g)

decode.strata(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\value{
\tabular{ll}{
  \code{decode} \tab a list with the results from \code{decode.strata}, \code{decode.locus.data}, and \code{decode.sequences}.\cr
  \code{decode.strata} \tab a named vector of strata designations.\cr
  \code{decode.locus.data} \tab a matrix of haplotypes or genotypes.\cr
  \code{decode.sequences} \tab a list of DNA sequences if present, otherwise \code{NULL}.\cr
}
}
\description{
Functions to decode \code{gtypes} objects.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

