% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diversity}
\alias{diversity}
\title{Unbiased Estimate of Diversity}
\usage{
diversity(x)
}
\arguments{
\item{x}{charcter or numeric vector or factor}
}
\description{
Calculate unbiased estimate of diversity for a vector of items
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

