% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dloop.g}
\alias{dloop.g}
\title{Dolphin dLoop gtypes Object}
\format{gtypes}
\usage{
data(dloop.g)
}
\description{
A \linkS4class{gtypes} object of 126 samples and 33 haplotypes.
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and 
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose 
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean. 
  Marine Mammal Science 31:1-20
}
\keyword{datasets}

