% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.seqs}
\alias{dolph.seqs}
\title{Dolphin mtDNA D-loop Sequences}
\format{list}
\usage{
data(dolph.seqs)
}
\description{
A list of 126 aligned control region sequences
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and 
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose 
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean. 
  Marine Mammal Science 31:1-20
}
\keyword{datasets}
