% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.strata}
\alias{dolph.strata}
\title{Dolphin Genetic Stratification and Haplotypes}
\format{data.frame}
\usage{
data(dolph.strata)
}
\description{
A data.frame of 126 samples with assignment of samples to either 
  broad-scale or fine-scale stratifications and mtDNA haplotype designations
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and 
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose 
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean. 
  Marine Mammal Science 31:1-20
}
\keyword{datasets}

