% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exptd.het}
\alias{exptd.het}
\title{Expected Heterozygosity}
\usage{
exptd.het(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\description{
Calculate expected heterozygosity for diploid data.
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5)
exptd.het(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{obsvd.het}}
}

