% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{fasta}
\alias{fasta}
\alias{read.fasta}
\alias{write.fasta}
\title{Read and Write FASTA}
\usage{
read.fasta(file)

write.fasta(x, file = "sequences.fasta")
}
\arguments{
\item{file}{a FASTA-formatted file of sequences.}

\item{x}{a list or a matrix of DNA sequences (see \code{\link[ape]{write.dna}}).}
}
\value{
for \code{read.fasta}, a set of sequences in DNAbin format.
}
\description{
Read and write FASTA formatted files of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

