% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\name{fastsimcoal}
\alias{fastsimcoal}
\title{Run fastsimcoal}
\usage{
fastsimcoal(num.pops, Ne, sample.size = NULL, sample.time = NULL,
  growth.rate = NULL, mig.rates = NULL, hist.ev = NULL,
  data.type = c("DNA", "MICROSAT", "SNP", "STANDARD", "FREQ"),
  locus.params = NULL, num.chrom = NULL, label = "strataG_fastsimcoal",
  num.sims = 1, inf.site.model = TRUE, quiet = TRUE,
  delete.files = TRUE, exec = "fsc252")
}
\arguments{
\item{num.pops}{number of populations.}

\item{Ne}{effective population size.}

\item{sample.size}{number of samples to take.}

\item{sample.time}{time to draw samples.}

\item{growth.rate}{growth rate of populations.}

\item{mig.rates}{list of migration matrices.}

\item{hist.ev}{matrix of historical events.}

\item{data.type}{the data type to be modeled. Can be one of
\code{"DNA", "MICROSAT", "SNP", "STANDARD", or "FREQ"}.}

\item{locus.params}{a list of locus parameter matrices with one element per
chromosome.}

\item{num.chrom}{number of chromosomes if multiple chromosomes with the same
structure are desired. Leave as \code{NULL} to take the number from the
number of elements in \code{locus.params}. If specified, the matrix in the
first element in \code{locus.params} will be used.}

\item{label}{character string to label files with.}

\item{num.sims}{number of simulations to run.}

\item{inf.site.model}{logical. Use infinite site model?}

\item{quiet}{logical. Run quietly?}

\item{delete.files}{logical. Delete files when done?}

\item{exec}{name of fastsimcoal executable.}
}
\value{
A list of \code{\link{gtypes}} objects for each simulated dataset.
}
\description{
Run a fastsimcoal simualtion and load results into a
  \linkS4class{gtypes} object.
}
\note{
Assumes that the program \code{fastsimcoal} is properly installed and
  available on the command line. On PC's, this requires having it in a
  folder in the PATH environmental variable. On Macs, the executable
  should be installed in a folder like \code{/usr/local/bin}. The actual name of
  the executable should be specified with the \code{exec} argument.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time
  coalescent simulator of genomic diversity under arbitrarily complex
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}

