% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixed.sites.R
\name{fixed.sites}
\alias{fixed.sites}
\title{Fixed Sites}
\usage{
fixed.sites(x, bases = c("a", "c", "g", "t", "-"))
}
\arguments{
\item{x}{a \code{\link{gtypes}} object with sequences, a list of sequences, or a consensus sequence.
Sequences must be aligned.}

\item{bases}{a character vector of valid bases to consider.}
}
\description{
Identify fixed sites among sequences.
}
\author{
Eric Archer <eric.archer@noaa.gov>
}
\seealso{
\code{\link{variable.sites}}
}

