% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc2gtypes.R
\name{fsc2gtypes}
\alias{fsc2gtypes}
\title{Load fastsimcoal output to gtypes}
\usage{
fsc2gtypes(p, sim = 1, type = NULL, gen.data = NULL, chrom = NULL,
  drop.mono = TRUE)
}
\arguments{
\item{p}{list of fastsimcoal input parameters and output.}

\item{sim}{number of the simulation replicate to read.}

\item{type}{type of marker to return.}

\item{gen.data}{matrix of parsed genetic data read from .arp file with 
`fscParseGeneticData()`.}

\item{chrom}{numerical vector giving chromosomes to return. `NULL` 
returns all chromosomes.}

\item{drop.mono}{drop monomorphic sites before creating `gtypes` object?}
}
\description{
Create a gtypes object from the result of a fastsimcoal run.
}
\note{
fastsimcoal is not included with `strataG` and must be downloaded 
  separately. Additionally, it must be installed such that it can be run from 
  the command line in the current working directory. See the vignette 
  for \code{external.programs} for installation instructions.
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time 
  coalescent simulator of genomic diversity under arbitrarily complex 
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  Excoffier, L., Dupanloup, I., Huerta-Sánchez, E., Sousa, V.C., 
  and M. Foll (2013) Robust demographic inference from genomic and SNP data. 
  PLOS Genetics, 9(10):e1003905. \cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}
\seealso{
\code{\link{fsc.input}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
