% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscRead.R
\name{fscRead}
\alias{fscRead}
\alias{fscReadArp}
\alias{fscReadParamEst}
\alias{fscReadSFSOutput}
\title{Read fastsimcoal output}
\usage{
fscReadArp(p, sim = 1, marker = c("all", "snp", "microsat", "dna",
  "standard"), chrom = NULL, sep.chrom = FALSE, drop.mono = FALSE,
  as.genotypes = TRUE, one.col = FALSE, sep = "/",
  coded.snps = FALSE)

fscReadParamEst(p)

fscReadSFSOutput(p)
}
\arguments{
\item{p}{list of fastsimcoal input parameters.}

\item{sim}{number of the simulation replicate to read.}

\item{marker}{type of marker to return.}

\item{chrom}{numerical vector giving chromosomes to return. `NULL` 
returns all chromosomes.}

\item{sep.chrom}{return a list with chromosomes separated?}

\item{drop.mono}{return only polymorphic loci?}

\item{as.genotypes}{return data as genotypes? If `FALSE`, original haploid 
data is returned. If `TRUE`, individuals are created by combining 
sequential haplotypes based on the ploidy used to run the simulation}

\item{one.col}{return genotypes with one column per locus? If `FALSE`, 
alleles are split into separate columns and designated as ".1", ".2", etc. 
for each locus.}

\item{sep}{character to use to separate alleles if `one.col = TRUE`.}

\item{coded.snps}{return diploid SNPs coded as 0 (major allele homozygote), 
1 (heterozygote), or 2 (minor allele homozygote). If this is `TRUE` and 
`marker = "snp"` and the data is diploid, genotypes will be returned with 
one column per locus.}
}
\value{
\describe{
 \item{fscReadArp}{Reads and parses Arlequin-formatted .arp output files 
   created by \code{fastsimcoal2}. Returns a data frame of genotypes, with 
   individuals created by combining haplotypes based on the stored value of
   ploidy specified when the simulation was run.}
 \item{fscReadParamEst}{Reads and parses files output from a 
   \code{fastsimcoal2} run conducted for parameter estimation. Returns a list 
   of data frames and vectors containing the data from each file.}
 \item{fscReadSFSOutput}{Reads site frequency spectra generated from 
   \code{fastsimcoal2}. Returns a list of the marginal and joint SFS.}
 }
}
\description{
Read arlequin formatted output or parameter estimation files 
  generated by fastsimcoal
}
\note{
fastsimcoal is not included with `strataG` and must be downloaded 
  separately. Additionally, it must be installed such that it can be run from 
  the command line in the current working directory. See the vignette 
  for \code{external.programs} for installation instructions.
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time 
  coalescent simulator of genomic diversity under arbitrarily complex 
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  Excoffier, L., Dupanloup, I., Huerta-Sánchez, E., Sousa, V.C., 
  and M. Foll (2013) Robust demographic inference from genomic and SNP data. 
  PLOS Genetics, 9(10):e1003905. \cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}
\seealso{
\code{\link{fsc.input}}, \code{\link{fscWrite}}, 
 \code{\link{fscRun}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
