% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscRun.R
\name{fscRun}
\alias{fscRun}
\alias{fscCleanup}
\title{Run fastsimcoal}
\usage{
fscRun(p, num.sims = 1, dna.to.snp = FALSE, max.snps = NULL,
  all.sites = TRUE, inf.sites = FALSE, sfs.type = c("daf", "maf"),
  num.ecm.loops = 20, num.cores = NULL, seed = NULL,
  exec = "fsc26")

fscCleanup(label, save.est = FALSE)
}
\arguments{
\item{p}{list of fastsimcoal input parameters and output.}

\item{num.sims}{number of simulation replicates to run.}

\item{dna.to.snp}{convert DNA sequences to numerical SNPs?}

\item{max.snps}{maximum number of SNPs to retain.}

\item{all.sites}{retain all sites? If \code{FALSE}, only polymorphic DNA 
sites will be returned. This includes SNP blocks as they are simulated as 
DNA sequences.}

\item{inf.sites}{use infinite sites model? If \code{TRUE}, all mutations are 
retained in the output, thus the number of sites for SNPs or DNA sequences 
will potentially be greater than what was requested.}

\item{sfs.type}{type of site frequency spectrum to compute for each 
population sample: `daf` = derived allele frequency (unfolded), 
`maf` = minor allele frequency (folded).}

\item{num.ecm.loops}{number of loops (ECM cycles) to be performed when 
estimating parameters from SFS. Default is 20.}

\item{num.cores}{number of cores to use.}

\item{seed}{random number seed for simulation.}

\item{exec}{name of fastsimcoal executable.}

\item{label}{character string to label files with.}

\item{save.est}{do not delete .est parameter estimation files during cleanup?}
}
\description{
Run a fastsimcoal simulation and load results into a 
  \linkS4class{gtypes} object.
}
\note{
fastsimcoal is not included with `strataG` and must be downloaded 
  separately. Additionally, it must be installed such that it can be run from 
  the command line in the current working directory. See the vignette 
  for \code{external.programs} for installation instructions.
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time 
  coalescent simulator of genomic diversity under arbitrarily complex 
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  Excoffier, L., Dupanloup, I., Huerta-Sánchez, E., Sousa, V.C., 
  and M. Foll (2013) Robust demographic inference from genomic and SNP data. 
  PLOS Genetics, 9(10):e1003905. \cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}
\seealso{
\code{\link{fsc.input}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
