% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstToNm.R
\name{fstToNm}
\alias{fstToNm}
\title{Nm from Fst}
\usage{
fstToNm(fst, ploidy)
}
\arguments{
\item{fst}{estimate of Fst between populations.}

\item{ploidy}{ploidy of locus Fst is from.}
}
\description{
Calculate Nm (number of migrants per generation) for a 
  given value of Fst.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

