% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gelato.plot}
\alias{gelato.plot}
\title{Plot GELATo}
\usage{
gelato.plot(gelato.result, unknown, main = NULL)
}
\arguments{
\item{gelato.result}{the result of a call to \code{\link{gelato.run}}.}

\item{unknown}{the name of an unknown stratum in the \code{x$likelihoods} element.}

\item{main}{main label for top of plot.}
}
\description{
Plot results from a GELATo run.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{gelato.run}}
}

