% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gtype.struct.func}
\alias{gtype.struct.func}
\alias{gtype.struct.stat}
\title{Population Structure Statistic Functions}
\description{
Functions that calculate population structure statistics.
The function takes a \code{\link{gtypes}} object and perhaps other parameters and returns
an object of class \code{gtype.struct.stat} which is a list containing:\cr
\tabular{ll}{
  \code{stat.name} \tab the name of the statistic.\cr
  \code{estimate} \tab the value of the statistic.\cr
  \code{strata.freq} \tab a \code{table} of the strata frequencies used in calculating the statistic.
This may be different from the frequencies in the original \code{gtypes} object due to exclusion of
samples due to missing data.\cr
}
}
\keyword{functions}

