% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypes.show.R
\name{gtypes.show}
\alias{gtypes.show}
\alias{show.gtypes}
\alias{.baseSmry}
\alias{.printSmryHeader}
\alias{.printBaseSmry}
\title{Show a gtypes object}
\usage{
\S4method{show}{gtypes}(object)

.baseSmry(g)

.printSmryHeader(x)

.printBaseSmry(x)
}
\arguments{
\item{object}{a \linkS4class{gtypes} object.}

\item{g}{a \linkS4class{gtypes} object.}

\item{x}{list from .baseSmry}
}
\description{
Show a gtypes object
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\keyword{internal}
