% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypes2df.R
\name{gtypes2df}
\alias{gtypes2df}
\title{Convert \code{gtypes} To \code{data.frame}}
\usage{
gtypes2df(g, one.col = TRUE, sep = "/", ...)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{one.col}{logical. If \code{TRUE}, then result has one column per 
locus.}

\item{sep}{character to use to separate alleles if \code{one.col} is 
\code{TRUE}.}

\item{...}{additional arguments ot be passed to or from methods.}
}
\value{
A \code{data.frame} with one row per sample.
}
\description{
Convert a \linkS4class{gtypes} object to a \code{data.frame}.
}
\author{
Eric Archer \email{eric.archer@noa.gov}
}
\seealso{
\link{df2gtypes}, \link{as.matrix.gtypes}
}

