% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypes2genlight.R
\name{gtypes2genlight}
\alias{gtypes2genlight}
\alias{genlight2gtypes}
\title{Convert Between \code{gtypes} And \code{genlight} objects.}
\usage{
gtypes2genlight(x)

genlight2gtypes(x)
}
\arguments{
\item{x}{either a \linkS4class{gtypes} or \linkS4class{genlight} object
to convert from.}
}
\description{
Convert a \code{gtypes} object to a \code{genlight} object 
  and vice-versa.
}
\examples{
data(msats.g)

# Create simple simulated SNPs
gl1 <- adegenet::glSim(n.ind = 100, n.snp.nonstruc = 1000, ploidy = 2)
gl1

# Convert to gtypes
gt <- genlight2gtypes(gl1)
gt

# Convert back to genlight
gl2 <- gtypes2genlight(gt)
gl2

}
\seealso{
\link{initialize.gtypes}, \link{df2gtypes}, 
  \link{sequence2gtypes}, \link{as.data.frame.gtypes}, 
  \link{gtypes2loci}, \link{gtypes2genind}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
