% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypesRF.R
\name{gtypesRF}
\alias{gtypesRF}
\title{gtype Random Forest}
\usage{
gtypesRF(g, gene = 1, pairwise = FALSE, conf.level = 0.95, unk = NULL, ...)
}
\arguments{
\item{g}{haploid \code{\link[strataG]{gtypes}} object with aligned sequences.}

\item{gene}{number or name of gene to use from multidna \code{@sequences} slot.}

\item{pairwise}{do analysis on all pairwise combinations of strata?}

\item{conf.level}{confidence level for the \code{\link{binom.test}} confidence interval}

\item{unk}{vector of strata to be treated as "unknowns" for prediction with}

\item{...}{arguments passed to \code{\link{sequenceRF}} and \code{\link[randomForest]{randomForest}}.}
}
\value{
a list containing a data.frame of summary statistics (\code{smry}), and the 
  \code{randomForest} object (\code{rf}). If \code{pairwise} is \code{TRUE} then 
  the \code{rf} element is a list of \code{randomForest} results for each row in
  \code{smry}.
}
\description{
Conduct Random Forest on a gtypes object.
}
\examples{
library(strataG)
data(dloop.g)

rf.result <- gtypesRF(dloop.g, pairwise = TRUE)

rf.result

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
