% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h.stats.R
\name{h.stats}
\alias{h.stats}
\title{Estimate Fixation Indices}
\usage{
h.stats(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\value{
matrix of Ho, Hs, and Ht for each locus.
}
\description{
Estimate Ho, Hs, and Ht from diploid data.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Nei, M. and R.K. Chesser. 1983. Estimation of fixation indices and gene diversities.
  Ann. Hum. Genet. 47:253-259.
}

