% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haplotypic.diversity.R
\name{haplotypic.diversity}
\alias{haplotypic.diversity}
\title{Haplotypic Diversity}
\usage{
haplotypic.diversity(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Calculate haplotypic diversity of all samples.
}
\note{
Function assumes that haplotypes in \code{g} have been already labelled. If not, use \code{\link{label.haplotypes}} first.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{diversity}}, \code{\link{nucleotide.diversity}}
}

