% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heterozygosity.R
\name{heterozygosity}
\alias{exptdHet}
\alias{heterozygosity}
\alias{obsvdHet}
\title{Observed and Expected Heterozygosity}
\usage{
exptdHet(g)

obsvdHet(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calculate observed heterozygosity for diploid data.
}
\note{
For a measure of haplotypic diversity (haploid "heterozygosity"),
  use \code{exptdHet}.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

exptdHet(msats)
obsvdHet(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

