% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.gtype.struct.func}
\alias{is.gtype.struct.func}
\title{Check Population Structure Function}
\usage{
is.gtype.struct.func(x)
}
\arguments{
\item{x}{an \code{R} object.}
}
\description{
Check if function is a valid population structure function. Used
  in overall and pairwise test functions.
}
\details{
checks if \code{x} is a \code{\link{gtypes}} population structure function.
  Used in permutation test functions.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

