% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.gtypes.R
\name{is.gtypes}
\alias{is.gtypes}
\title{Test if object is \code{gtypes}}
\usage{
is.gtypes(x)
}
\arguments{
\item{x}{R object to be tested.}
}
\value{
Logical stating if 'x' is a \code{\linkS4class{gtypes}} object.
}
\description{
Test if object is \code{gtypes}
}
\examples{
data(msats.g)
is.gtypes(msats.g) # TRUE

data(dolph.msats)
is.gtypes(dolph.msats) # FALSE

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

