% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape2gtypes.R
\name{landscape2gtypes}
\alias{landscape2gtypes}
\title{Create gtypes from Rmetasim landscape}
\usage{
landscape2gtypes(Rland)
}
\arguments{
\item{Rland}{rmetasim landscape object}
}
\description{
Create a gtypes object from an Rmetasim landscape object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
