% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linkage.genepop}
\alias{linkage.genepop}
\title{Linkage Disequlibrium}
\usage{
linkage.genepop(g, show.output = TRUE, delete.files = TRUE,
  label = "linkage.genepop", ...)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{show.output}{logical. Show GENEPOP output on console?}

\item{delete.files}{logical. Delete GENEPOP input and output files when done?}

\item{label}{character string to use to label GENEPOP input and output files.}

\item{...}{other arguments to be passed to \code{\link{genepop.run}}.}
}
\description{
Calculate linkage disequilibrium p-values using GENEPOP.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

