% Generated by roxygen2 (4.0.2): do not edit by hand
\name{low.freq.subs}
\alias{low.freq.subs}
\title{Low Frequency Substitutions}
\usage{
low.freq.subs(dna.seq, min.freq = 3, motif.length = 10)
}
\arguments{
\item{dna.seq}{a list of DNA sequences.}

\item{min.freq}{minimum frequency of base to be flagged.}

\item{motif.length}{length of motif around low frequency base to output.}
}
\value{
data.frame listing id, site number, and motif around low frequency base call.
}
\description{
Check nucleotide sites for low frequency substitutions.
}
\examples{
data(dolph.haps)
low.freq.subs(dolph.haps)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

