% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{Minimum Allele Frequencies}
\usage{
maf(g, by.strata = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{logical - return results grouped by strata?}
}
\value{
A vector or matrix of minimum allele frequencies at each locus.
}
\description{
Calculate minimum allele frequencies for each locus.
}
\examples{
data(msats.g)

maf(msats.g)

}
\seealso{
\link{alleleFreqs}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
