% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge.gtypes}
\alias{merge.gtypes}
\title{Merge gtypes}
\usage{
\method{merge}{gtypes}(x, y, description, ...)
}
\arguments{
\item{x,y}{two \code{\link{gtypes}} objects.}

\item{description}{optional description for resulting object. If \code{NULL} then descriptions of both objects are combined.}

\item{...}{other parameters (ignored).}
}
\description{
Merge two conforming \code{\link{gtypes}} into one.
}
\details{
\code{x} and \code{y} must have the same ploidy. If haploid, they cannot share samples. If diploid, and shared samples
  cannot also share loci in the two objects.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

