% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microhaplot2rubias.R
\name{microhaplot2rubias}
\alias{microhaplot2rubias}
\title{Convert MicroHaplot file to rubias format}
\usage{
microhaplot2rubias(x, sample.type = c("reference", "mixture"))
}
\arguments{
\item{x}{data frame from MicroHaplot or the name a of .csv file.}

\item{sample.type}{determines type of MicroHaplot sample. Can be any 
unambiguous substring of \code{"reference"} or \code{"mixture"}.}
}
\description{
Converts a reference or mixture file from the MicroHaplot 
  package to the proper format for the rubias package
}
\author{
Eric Archer \email{eric.archer@noaa.gov} Converted from code by
  John Horne (\url{https://johnbhorne.wordpress.com/2019/03/22/snp-haplotypes-for-mixed-stock-fishery-analysis-microhaplot-to-rubias-conversion-in-r/})
}
