% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/most.distant.sequences.R
\name{most.distant.sequences}
\alias{most.distant.sequences}
\title{Most Distant Sequences}
\usage{
most.distant.sequences(x, num.seqs = NULL, model = "raw",
  pairwise.deletion = TRUE)
}
\arguments{
\item{x}{a \code{\link{gtypes}} object with aligned sequences, or a list of aligned DNA sequences.}

\item{num.seqs}{number of sequences to return.}

\item{model}{a character string specifying the evolutionary model to be used.
See \link{dist.dna} for more information.}

\item{pairwise.deletion}{a logical indicating whether to delete sites with missing data.
  See \link{dist.dna} for more information.}
}
\value{
a vector of the sequence names that are have the greatest mean pairwise distance and
  smallest variance of pairwise distances. The names are returned in order from most to least distant.
}
\description{
Finds the set of sequences that have the greatest mean pairwise distance and
  smallest variance of pairwise distances.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

