% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.omit.gtypes.R
\name{na.omit.gtypes}
\alias{na.omit.gtypes}
\title{Omit Missing Data}
\usage{
\method{na.omit}{gtypes}(object, strata = TRUE, loci = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{gtypes}} object.}

\item{strata}{logical. Remove unstratified samples?}

\item{loci}{logical. Remove samples with missing data for all loci?}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Omit samples or loci with missing data in \code{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

