% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nei.Da.R
\name{nei.Da}
\alias{Da}
\alias{nei.Da}
\title{Nei's Da}
\usage{
nei.Da(g, na.rm = FALSE)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{na.rm}{logical. Delete loci which have missing data for every sample in a stratum?}
}
\description{
Calcuate frequency-based Nei's Da for haploid or diploid data.
}
\details{
Returns Nei's Da for each pair of strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Nei et al 1983 Accuracy of Estimated Phylogenetic Trees from
  Molecular Data. J Mol Evol 19:153-170 (eqn 7)\cr
  Nei, M., and S. Kumar (2000) Molecular Evolution and Phylogenetics.
  Oxford University Press, Oxford. (pp. 268, eqn 13.6)
}

