% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neiDa.R
\name{neiDa}
\alias{neiDa}
\title{Nei's Da}
\usage{
neiDa(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calcuate frequency-based Nei's Da for haploid or diploid data.
}
\details{
Returns Nei's Da for each pair of strata.
}
\examples{
data(msats.g)

neiDa(msats.g)
 
}
\references{
Nei et al 1983 Accuracy of Estimated Phylogenetic Trees from 
  Molecular Data. J Mol Evol 19:153-170 (eqn 7)\cr
  Nei, M., and S. Kumar (2000) Molecular Evolution and Phylogenetics. 
  Oxford University Press, Oxford. (pp. 268, eqn 13.6)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
