% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nucleotide.diversity}
\alias{nucleotide.diversity}
\title{Nucleotide Diversity}
\usage{
nucleotide.diversity(x, ignore.gaps = TRUE)
}
\arguments{
\item{x}{a list of sequences or a \code{\link{gtypes}} object.}

\item{ignore.gaps}{logical. Ignore gaps when calculating diversity?}
}
\value{
Nucleotide diversity by site.
}
\description{
Calculate nucleotide diversity for set of haplotypes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

