% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numMissing.R
\name{numMissing}
\alias{numMissing}
\title{Number Missing Data}
\usage{
numMissing(g, prop = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{prop}{logical determining whether to return proportion missing.}
}
\value{
a vector of loci with number (or, if \code{prop = TRUE},
  the proportion) of individuals missing data for at least one allele.
}
\description{
Calculate the number of individuals with missing data by locus.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

numMissing(msats)
numMissing(msats, prop = TRUE)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

