% Generated by roxygen2 (4.0.2): do not edit by hand
\name{obsvd.het}
\alias{obsvd.het}
\title{Observed Heterozygosity}
\usage{
obsvd.het(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Calculate observed heterozygosity for diploid data.
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5)
obsvd.het(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{exptd.het}}
}

