% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p.val.R
\name{p.val}
\alias{p.val}
\title{Permutation Test P-value}
\usage{
p.val(obs, null.dist)
}
\arguments{
\item{obs}{observed value.}

\item{null.dist}{vector of values from permutation null distribution.}
}
\description{
Calculate the p-value for a permutation test.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

