% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLocusNames.R
\name{parseLocusNames}
\alias{.parseLocusNames}
\alias{parseLocusNames}
\title{Parse Locus Names}
\usage{
.parseLocusNames(locus.names, ploidy)
}
\arguments{
\item{locus.names}{a vector of column names, where each locus must be
named with the same roots. For example, diploid locus 'ABCD' would have
two columns named something like 'ABCD.1' and 'ABCD.2', or
'ABCD_A' and 'ABCD_B'.\cr}

\item{ploidy}{integer representing the ploidy of the data.}
}
\value{
A vector of the root locus names.
}
\description{
Parse locus names from a matrix where alleles are in
  consecutive columns.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

