% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pct.unique.haplotypes.R
\name{pct.unique.haplotypes}
\alias{pct.unique.haplotypes}
\title{Percent Unique Haplotypes}
\usage{
pct.unique.haplotypes(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Calculate the percent of haplotypes that are unique.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

