% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteStrata.R
\name{permuteStrata}
\alias{permuteStrata}
\title{Permute strata}
\usage{
permuteStrata(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a \linkS4class{gtypes} object with the strata randomly permuted.
}
\description{
Permute the strata slot within a \linkS4class{gtypes} object.
}
\examples{
data(msats.g)
msats.g <- stratify(msats.g, "fine")
summary(msats.g)

ran.msats <- permuteStrata(msats.g)
summary(ran.msats)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

