% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hstats.R, R/popStructStat.R, R/statChi2.R, R/statFis.R, R/statFst.R, R/statGst.R, R/statJostD.R, R/statPhist.R
\name{Hstats}
\alias{Hstats}
\alias{popStructStat}
\alias{statChi2}
\alias{statFis}
\alias{statFst}
\alias{statFstPrime}
\alias{statGst}
\alias{statGstDblPrime}
\alias{statGstPrime}
\alias{statJostD}
\alias{statPhist}
\title{Population structure statistics}
\usage{
Hstats(g)

statChi2(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statFis(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statFst(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statFstPrime(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statGst(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statGstPrime(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE,
  prime.type = c("nei", "hedrick"), ...)

statGstDblPrime(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statJostD(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE, ...)

statPhist(g, nrep = NULL, strata.mat = NULL, keep.null = FALSE,
  model = "K80", gamma = FALSE, pairwise.deletion = TRUE, ...)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{nrep}{number specifying number of permutation replicates to use for 
permutation test.}

\item{strata.mat}{an optional matrix of permuted stratifications. If given, 
the first column is taken to be the observed stratification and all others 
are random permutations. Strata must be defined by sequential integer values 
starting with 0. If \code{NULL} stratification is taken from \code{g}. This 
argument is primarily used internally in \code{\link{overallTest}}.}

\item{keep.null}{logical. Keep the null distribution from the 
permutation test?}

\item{...}{optional arguments passed to or from other functions.}

\item{prime.type}{type of G'st to calculate. Can be "nei" or "hedrick".}

\item{model, gamma, pairwise.deletion}{parameters passed to 
\code{\link[ape]{dist.dna}}.}
}
\value{
the named statistic estimate.
}
\description{
Population structure statistics
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

