% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.gtypeSummary}
\alias{print.gtypeSummary}
\title{Print gtypeSummary}
\usage{
\method{print}{gtypeSummary}(x, ... )
}
\arguments{
\item{x}{list from summary.gtypes}

\item{...}{other arguments (ignored)}
}
\description{
Print the output for a \code{gtypeSummary} object from \code{\link{summary.gtypes}}.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

