% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/privateAlleles.R
\name{privateAlleles}
\alias{privateAlleles}
\title{Private Alleles}
\usage{
privateAlleles(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
matrix with the number of private alleles in each strata at each locus.
}
\description{
The number of private alleles in each strata and locus.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

privateAlleles(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

