% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual2prob.R
\name{qual2prob}
\alias{qual2prob}
\title{Convert FASTQ quality scores to base probability}
\usage{
qual2prob(x, phred.scale = c("+33", "+64"), simplify = TRUE)
}
\arguments{
\item{x}{a vector of quality scores represented as ASCII characters from 
\code{!} to \code{~}. Each element in \code{x} can be a string of 
characters or a single character.}

\item{phred.scale}{scale of PHRED encoding. Can be : 
\code{"+33"} = range of \code{0:93} (e.g., Sanger, Illumina 1.8+), or
\code{"+64"} = range of \code{-31:62} (e.g., Solexa, Illumina 1.3+, 1.5+)}

\item{simplify}{if there is only one set of quality scores, simplify 
resulting list to a vector?}
}
\value{
List of probabilities that each base represented by the characters in
  \code{x} was correctly called. List has one element for each set of 
  quality scores in \code{x}.
}
\description{
Converts FASTQ quality scores encoded as ASCII characters to
  numeric probability that base was correctly called.
}
\examples{
# each element is a single quality score
quality.1 <- sapply(sample(33:126, 15, rep = TRUE), intToUtf8)
prob.1 <- qual2prob(quality.1)

quality.1
prob.1

# each element is a string of quality scores
quality.2 <- replicate(5, {
  num.sites <- sample(5:10, 1)
  chars <- sapply(sample(33:126, num.sites, rep = TRUE), intToUtf8)
  paste(chars, collapse = "")
}, simplify = "vector")
prob.2 <- qual2prob(quality.2)

quality.2
prob.2 

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
