% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.arlequin.r, R/write.arlequin.R
\name{read.arlequin}
\alias{arlequin}
\alias{read.arlequin}
\alias{write.arlequin}
\title{Read and Write Arlequin Files}
\usage{
read.arlequin(file)

write.arlequin(g, file = "gtypes.prj", title = "gtypes from R",
  data.type = "DNA")
}
\arguments{
\item{file}{filename for output file.}

\item{g}{a \code{\link{gtypes}} object.}

\item{title}{title for data in file.}

\item{data.type}{type of data. Can be "DNA", "RFLP", or "MICROSAT".}
}
\description{
Read and write an Arlequin-formatted files.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Excoffier, L. G. Laval, and S. Schneider (2005) Arlequin ver. 3.0: An integrated
  software package for population genetics data analysis. Evolutionary Bioinformatics Online 1:47-50.\cr
  Arlequin available at \url{http://cmpg.unibe.ch/software/arlequin3/}
}

