% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.fasta}
\alias{fasta}
\alias{read.fasta}
\alias{write.fasta}
\title{Read and Write FASTA}
\usage{
read.fasta(file)

write.fasta(x, file = "sequences.fasta")
}
\arguments{
\item{file}{a FASTA-formatted file of sequences.}

\item{x}{a list of DNA sequences or a haploid \code{\link{gtypes}} object with sequences.}
}
\value{
from \code{read.fasta}, a list of DNA sequences.
}
\description{
Read and write FASTA formatted files of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{as.dna.seq}}
}

