% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.gen.data}
\alias{read.gen.data}
\title{Read Genetic Data}
\usage{
read.gen.data(file, na.strings = c(NA, "NA", "", " ", "?", "."), ...)
}
\arguments{
\item{file}{filename of .csv file.}

\item{na.strings}{see \code{\link{read.table}}.}

\item{...}{other arguments passed to \code{\link{read.table}}.}
}
\value{
a \code{data.frame}.
}
\description{
Reads a .csv file into a \code{data.frame} formatted as genetic data. Replaces commonly
  used missing data values with NA and removes blank lines.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

