% Generated by roxygen2 (4.0.2): do not edit by hand
\name{relabel.sequences}
\alias{relabel.sequences}
\title{Relabel Sequences}
\usage{
relabel.sequences(g, new.labels, add.labels = TRUE, sep = ".")
}
\arguments{
\item{g}{\code{\link{gtypes}} object.}

\item{new.labels}{character vector giving new labels for sequences. Must either be of length 1 or the
number of sequences in \code{g}. If of length 1 and \code{add.labels} = FALSE, sequences will be
sequentially numbered.}

\item{add.labels}{logical. Should \code{new.labels} be added to the existing sequence labels?}

\item{sep}{character separator between new and old labels (if \code{add.labels} = FALSE), or new
  label and sequential numbers.}
}
\value{
a \code{gtypes} object with sequences relabelled.
}
\description{
Relabel sequences in a \code{gtypes} object by changing names or adding to names.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{as.dna.seq}}
}

