% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSequences.R
\name{removeSequences}
\alias{removeSequences}
\title{Remove Sequences}
\usage{
removeSequences(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a new \linkS4class{gtypes} object with unused sequences removed.
}
\description{
Remove sequences not used by samples listed in \code{@loci} 
  slot.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

