% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUnusedSequences.R
\name{removeUnusedSequences}
\alias{removeUnusedSequences}
\title{Remove Unused Sequences}
\usage{
removeUnusedSequences(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a new \linkS4class{gtypes} object with unused sequences removed.
}
\description{
Remove sequences not used by samples.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
