% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence2gtypes.R
\name{sequence2gtypes}
\alias{sequence2gtypes}
\title{Convert Sequences To \code{gtypes}}
\usage{
sequence2gtypes(x, strata = NULL, seq.names = NULL, description = NULL,
  other = NULL)
}
\arguments{
\item{x}{DNA sequences as a character matrix, a \code{\link{DNAbin}} object,
or \linkS4class{multidna} object.}

\item{strata}{a vector or factor giving stratification for each sequence.}

\item{seq.names}{names for each set of sequences.}

\item{description}{a label for the object (optional).}

\item{other}{a slot to carry other related information - unused in package
  analyses (optional).}
}
\value{
a \linkS4class{gtypes} object.
}
\description{
Load sequence data into a \linkS4class{gtypes} object.
}
\examples{
#--- create a haploid sequence (mtDNA) gtypes object
data(dolph.strata)
data(dolph.seqs)
strata <- dolph.strata$fine
names(strata) <- dolph.strata$ids
dloop.fine <- sequence2gtypes(dolph.seqs, strata, seq.names = "dLoop",
  description = "dLoop: fine-scale stratification")
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

