% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRF.R
\name{sequenceRF}
\alias{sequenceRF}
\title{Sequence Random Forest}
\usage{
sequenceRF(
  x,
  replace = FALSE,
  sampsize = NULL,
  train.pct = 0.5,
  min.n = 2,
  nrep = 0,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{data.frame of stratified and aligned sequences from \code{\link{gtypes2rfDF}}.}

\item{replace}{sample with replacement? passed to \code{\link[randomForest]{randomForest}}.}

\item{sampsize}{number of samples to take from each strata. passed to \code{\link[randomForest]{randomForest}}.
If \code{NULL}, value is set to the minumum sample size * \code{train.pct}.}

\item{train.pct}{proportion of strata to use for training if sampsize is \code{NULL}.}

\item{min.n}{minimum sample size to use. If sample size for any strata is less this, 
this value will be used instead.}

\item{nrep}{number of rfPermute replicates.}

\item{conf.level}{confidence level for the \code{\link{binom.test}} confidence interval}

\item{...}{arguments passed to \code{\link[randomForest]{randomForest}}.}
}
\value{
a list containing a data.frame of summary statistics (\code{smry}), 
  and the \code{rfPermute} object (\code{rp}).
}
\description{
Conduct Random Forest on stratified data.frame of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
