% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shared.haps.R
\name{shared.haps}
\alias{shared.haps}
\title{Shared Haplotypes}
\usage{
shared.haps(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\value{
a data.frame identifying the hapltoypes shared between each pair of strata.
  If no haplotypes are shared, returns \code{NA}.
}
\description{
Identify haplotypes shared between strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

