% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharedLoci.R
\name{sharedLoci}
\alias{propSharedIds}
\alias{propSharedLoci}
\alias{sharedAlleles}
\alias{sharedLoci}
\title{Shared Loci}
\usage{
propSharedLoci(g, type = c("strata", "ids"), num.cores = 1)

sharedAlleles(g, smry = "num")

propSharedIds(ids, g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{type}{a character vector determining type of pairwise comparsion. Can
be "strata" for strata or "ids" for individuals.}

\item{num.cores}{number of CPU cores to use.}

\item{smry}{a character vector determining type of summary for
\code{sharedAlleles}. "which" returns the names of the alleles shared.
"num" returns the number of alleles shared.}

\item{ids}{character vector of two sample ids to compare.}
}
\value{
data.frame summary of pairwise shared loci.
}
\description{
Calculate proportion of alleles and number of loci shared
  between pairs of individuals or strata.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

propSharedLoci(msats)

sharedAlleles(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

