% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.gtypes.R
\docType{methods}
\name{show,gtypes-method}
\alias{show,gtypes-method}
\alias{show.gtypes}
\title{Show a gtypes object}
\usage{
\S4method{show}{gtypes}(object)
}
\arguments{
\item{object}{a \linkS4class{gtypes} object.}
}
\description{
Show a gtypes object
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

