% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smry.by.locus.R
\name{smry.by.locus}
\alias{smry.by.locus}
\title{By-Locus Summaries}
\usage{
smry.by.locus(g, label)
}
\arguments{
\item{g}{a \code{\link[strataG]{gtypes}} object.}

\item{label}{label for output folder and prefix for files.}
}
\description{
By-Locus Summaries
}
\note{
Requires that GENEPOP is installed on the system and accessible from the command line.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

