% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smry.by.sample.R
\name{smry.by.sample}
\alias{smry.by.sample}
\title{By-Sample Summaries}
\usage{
smry.by.sample(g, label)
}
\arguments{
\item{g}{a \code{\link[strataG]{gtypes}} object.}

\item{label}{label for output folder and prefix for files.}
}
\description{
By-Sample Summaries
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

