% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat.chi2.R
\name{stat.chi2}
\alias{stat.chi2}
\title{CHI-squared Analysis of Population Structure}
\usage{
stat.chi2(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\value{
a \code{\link{gtype.struct.stat}} list with the CHI-squared estimate.
}
\description{
CHI-squared Analysis of Population Structure
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

stat.chi2(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

