% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stat.fst.prime}
\alias{stat.fst.prime}
\title{F'st analysis of population structure}
\usage{
stat.fst.prime(g)
}
\arguments{
\item{g}{a \code{gtypes} object}
}
\description{
F'st analysis of population structure
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

stat.fst.prime(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

