% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat.gst.prime.hedrick.R
\name{stat.gst.prime.hedrick}
\alias{stat.gst.prime.hedrick}
\title{Hedrick's G'st Analysis of Population Structure}
\usage{
stat.gst.prime.hedrick(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\description{
Hedrick's G'st Analysis of Population Structure
}
\details{
Calculate G'st from Hedrick 2005
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

stat.gst.prime.hedrick(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

