% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strata.pairs}
\alias{strata.pairs}
\title{Strata Pairs}
\usage{
strata.pairs(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Make a data.frame of all pairs of strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

