% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strata.split.R
\name{strata.split}
\alias{strata.split}
\title{Split Strata}
\usage{
strata.split(g, strata = NULL, keep.unstratified = FALSE, ...)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{strata}{a character vector giving a subset of strata to select. If \code{NULL} then a list with
all strata is created.}

\item{keep.unstratified}{logical. Create a gtypes object with unstratified individuals?}

\item{...}{other parameters to be passed to \code{\link{subset.gtypes}}}
}
\value{
A named list where each element is a \code{gtypes} object for a single stratum in \code{g}.
}
\description{
Return a list of gtypes for each strata
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{subset.gtypes}} to create a single \code{gtypes} object for specific strata.
}

