% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataGUI.R
\name{strataGUI}
\alias{strataGUI}
\title{strataG GUI}
\usage{
strataGUI()
}
\description{
A graphical user interface for creating gtypes objects, conducting 
  quality control analyses, and analyses of population structure.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
