% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataSplit.R
\name{strataSplit}
\alias{strataSplit}
\title{Split Strata}
\usage{
strataSplit(g, strata = NULL, remove.sequences = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{strata}{a character vector giving a subset of strata to select. 
If \code{NULL} then a list with all strata is created.}

\item{remove.sequences}{logical. If \code{TRUE} any sequences not referenced 
in selected samples will not be in the returned object.}
}
\value{
A named list where each element is a \code{gtypes} object 
  for a single stratum in \code{g}.
}
\description{
Return a list of \code{gtypes} for each stratum.
}
\examples{
data(msats.g)

# Proportion of unique alleles in each stratum
msats.list <- strataSplit(msats.g)
lapply(msats.list, propUniqueAlleles)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
