% Generated by roxygen2 (4.0.2): do not edit by hand
\name{structure.parse.q.mat}
\alias{structure.parse.q.mat}
\title{Parse Q-Matrix Text}
\usage{
structure.parse.q.mat(q.mat.txt, pops)
}
\arguments{
\item{q.mat.txt}{character vector of Q matrix from STRUCTURE output file.}

\item{pops}{vector of population labels to be used in place of numbers in STRUCTURE file.}
}
\description{
Parse Q-Matrix Text
}
\details{
Function is used internally by \code{\link{structure.read}} and \code{\link{clumpp.run}} to
  parse output files.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

