% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subset.gtypes}
\alias{subset.gtypes}
\title{Subset a gtypes Object}
\usage{
\method{subset}{gtypes}(x, ids, strata, loci, remove.sequences, ...)
}
\arguments{
\item{x}{a \code{\link{gtypes}} object.}

\item{ids}{a character vector specifying which ids to keep.}

\item{strata}{a character vector specifying which strata to keep.}

\item{loci}{a character vector specifying which loci to keep.}

\item{remove.sequences}{logical. For haploid objects, remove sequences for samples that are excluded?}

\item{...}{other parameters (ignored).}
}
\description{
Return a subset of specific individuals, strata, or loci from a gtypes object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

