% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarize.sequences.R
\name{summarize.sequences}
\alias{summarize.sequences}
\title{DNA Sequence Summaries}
\usage{
summarize.sequences(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\value{
A data.frame with sequence summary statistics with one row per stratum and a final row
  for the all sequences in \code{g}.
}
\description{
Generate a standard set of summaries for sequences.
}
\note{
The function labels haplotypes first so may run slower for large data sets with individual sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{haplotypic.diversity}}, \code{\link{variable.sites}}
}

