% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSamples.R
\name{summarizeInd}
\alias{summarizeInd}
\title{Individual Summaries}
\usage{
summarizeInd(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
A data.frame with rows for each sample and columns containing:
\describe{
  \item{\code{id}}{The individual id}
  \item{\code{stratum}}{The stratum of the individual}
  \item{\code{num.loci.missing.genotypes}}{The number of genotypes missing}
  \item{\code{pct.loci.missing.genotypes}}{The proportion of genotypes missing}
  \item{\code{pct.loci.homozygous}}{The proportion of loci homozygous}
}
}
\description{
Compile standard by-individual summaries.
}
\examples{
data(msats.g)

summarizeSamples(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
