% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSeqs.R
\name{summarizeSeqs}
\alias{summarizeSeqs}
\title{Sequence Summaries}
\usage{
summarizeSeqs(x)
}
\arguments{
\item{x}{a \code{\link[ape]{DNAbin}} object.}
}
\value{
a matrix listing the start and end positions of each sequence 
  (excluding beginning and trailing N's), the length, the number of N's, 
  and the number of indels.
}
\description{
Summaries for each sequence.
}
\examples{
library(apex)
data(woodmouse)

summarizeSeqs(woodmouse)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
