% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gtypes.R
\docType{methods}
\name{summary,gtypes-method}
\alias{print.gtypeSummary}
\alias{summary,gtypes-method}
\title{Summarize gtypes Object}
\usage{
\S4method{summary}{gtypes}(object, ...)

\method{print}{gtypeSummary}(x, ...)
}
\arguments{
\item{object}{a \linkS4class{gtypes} object.}

\item{...}{other arguments (ignored).}

\item{x}{list from summary.gtypes}
}
\value{
a list with the following elements:
\tabular{ll}{
  \code{num.ind} \tab number of individuals.\cr
  \code{num.loc} \tab number of loci.\cr
  \code{num.strata} \tab number of strata.\cr
  \code{allele.freqs} \tab a list with tables of allele frequencies 
    by strata.\cr
  \code{strata.smry} \tab a by-strata data.frame summarizing haplotypes 
    or loci.\cr
  \code{locus.smry} \tab a data.frame summarizing each locus for 
    non-haploid objects, \code{NULL} for haploid objects.\cr
  \code{seq.smry} \tab a summary of the sequence length and base 
    frequencies.\cr
}
}
\description{
Generate a summary of a \code{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

