% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gtypes}
\alias{summary.gtypes}
\title{Summarize gtypes Object}
\usage{
\method{summary}{gtypes}(object, ... )
}
\arguments{
\item{object}{a \code{\link{gtypes}} object.}

\item{...}{other arguments (ignored).}
}
\value{
a list with the following elements:
\tabular{ll}{
  \code{all} \tab vector of number of samples, strata, and loci.\cr
  \code{locus.freq} \tab for haploid \code{gtypes}, a frequency table of haplotypes by strata, and for diploid.\cr
  \code{gtypes} \tab a list of strata containing a list of data.frames of allele frequencies for each loci.\cr
  \code{by.strta} \tab a by-strata data.frame summarizing haplotypes or loci.\cr
  \code{num.seq} \tab if haploid, the number of sequences, otherwise \code{NA}.\cr
}
}
\description{
Generate a summary of a \code{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

