% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tajimas.d}
\alias{tajimas.d}
\title{Tajima's D}
\usage{
tajimas.d(x)
}
\arguments{
\item{x}{a FASTA filename, list of aligned DNA sequences,
  or a haploid \code{\link{gtypes}} object with sequences}
}
\value{
A named vector with the estimate for \code{D} and the \code{p.value} that it is different from 0.
}
\description{
Calculate Tajima's D for a set of sequences to test for selection.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Tajima, F. 1989. Statistical method for testing the neutral mutation
  hypothesis by DNA polymorphism. Genetics 123:585-595.\cr
  Fu, Y-X. 1997. Statistical tests of neutrality of mutations against population growth, hitchiking
  and background selection. Genetics 147:915-925.
}

