% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiTvRatio.R
\name{TiTvRatio}
\alias{TiTvRatio}
\alias{isTi}
\alias{isTv}
\alias{subType}
\title{Transition / Transversion Ratio}
\usage{
TiTvRatio(x)

subType(b1, b2)

isTi(b1, b2)

isTv(b1, b2)
}
\arguments{
\item{x}{sequence of DNA.}

\item{b1, b2}{two bases to be compared.}
}
\value{
For \code{TiTvRatio}, a vector providing: \cr
\tabular{ll}{
  \code{Ti} \tab the number of transitions.\cr
  \code{Tv} \tab the number of transversions.\cr
  \code{Ti.Tv.ratio} \tab the transition/transversion ratio.\cr
}
For \code{subType} either "ti" for transition, or "tv" for transversion.\cr
For \code{isTi} and \code{isTv} a logical identifying whether 
  the \code{b1} to \code{b2} is a transition or transversion.
}
\description{
Calculate transition/transversion ratio. Test substitution 
  type of two bases.
}
\examples{
data(dolph.seqs)

TiTvRatio(dolph.seqs)

subType("a", "c")

isTi("a", "c")

isTv("a", "c")

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

