% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trim.ns.R
\name{trim.ns}
\alias{trim.ns}
\title{Trim N's From Sequences}
\usage{
trim.ns(dna.seq)
}
\arguments{
\item{dna.seq}{a list of DNA sequences.}
}
\value{
a list of DNA sequences.
}
\description{
Removes N's from beginning and end of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

