% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.iupac.codes.R
\name{valid.iupac.codes}
\alias{valid.iupac.codes}
\title{Valid IUPAC Codes}
\usage{
valid.iupac.codes(bases)
}
\arguments{
\item{bases}{character vector of nucleotides or IUPAC codes to be checked.}
}
\value{
character vector of valid IUPAC codes for \code{bases}.
}
\description{
Get all possible valid IUPAC codes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

