% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wright.fst}
\alias{wright.fst}
\title{Wright's Fst}
\usage{
wright.fst(Ne, dispersal, gen.time)
}
\arguments{
\item{Ne}{Effective population size.}

\item{dispersal}{migration rate in terms of probability of an individual migrating in a generation.}

\item{gen.time}{number of generations since ancestral population.}
}
\description{
Calcualte Wright's Fst from Ne, dispersal, and generation time.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

