% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrightFst.R
\name{wrightFst}
\alias{wrightFst}
\title{Wright's Fst}
\usage{
wrightFst(Ne, dispersal, gen.time, ploidy)
}
\arguments{
\item{Ne}{Effective population size.}

\item{dispersal}{migration rate in terms of probability of an individual 
migrating in a generation.}

\item{gen.time}{number of generations since ancestral population.}

\item{ploidy}{ploidy of the locus}
}
\description{
Calcualte Wright's Fst from Ne, dispersal, and generation time.
}
\examples{
dispersal <- seq(0.05, 0.8, by = 0.05)
fst <- wrightFst(100, dispersal, 20, 2)
plot(dispersal, fst, type = "l")

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\link{numGensEq}, \link{fstToNm}, \link{expectedNumAlleles}
}

