% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.arlequin.R
\name{write.arlequin}
\alias{arlequin}
\alias{write.arlequin}
\title{Write Arlequin Input Files}
\usage{
write.arlequin(g, label = NULL, locus = 1)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{label}{label for Arlequin project filename (.arp). If \code{NULL}, the 
gtypes description is used if present.}

\item{locus}{numeric or character designation of which locus to write for 
haploid data.}
}
\description{
Create an Arlequin-formatted input file from a \code{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Excoffier, L.G. Laval, and S. Schneider (2005) 
  Arlequin ver. 3.0: An integrated software package for population genetics 
  data analysis. Evolutionary Bioinformatics Online 1:47-50.\cr
  Available at \url{http://cmpg.unibe.ch/software/arlequin3/}
}

