% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.nexus.snapp.R
\name{write.nexus.snapp}
\alias{write.nexus.snapp}
\title{Write NEXUS File for SNAPP}
\usage{
write.nexus.snapp(g, file = "snapp.data.nex")
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{file}{the filename the NEXUS file to output.}
}
\description{
Write NEXUS File for SNAPP
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
