% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygosity.R
\name{zygosity}
\alias{zygosity}
\title{Identify zygosity across loci}
\usage{
zygosity(x)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}
}
\value{
data frame that identifies if the genotype for an individual
  (\code{id}) at a locus (\code{locus}) is a heterozygote (\code{het}), 
  homozygote (\code{hom}) or missing (\code{NA}). Note that if haploid, 
  all non-missing values will be \code{hom}.
}
\description{
Identifies genotypes as heterozygous or homozygous.
}
\examples{
data(msats.g)

msats.zyg <- zygosity(msats.g)
head(msats.zyg, 10)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
