% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicRichness.R
\name{allelicRichness}
\alias{allelicRichness}
\title{Allelic Richness}
\usage{
allelicRichness(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
the allelic richness of each locus calculated as the number of 
  alleles divided by the number of samples without missing data at 
  that locus.
}
\description{
Calculate allelic richness for each locus.
}
\examples{
data(msats.g)
allelicRichness(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

