% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.gtypes}
\alias{as.data.frame.gtypes}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{gtypes}(x, ...)
}
\arguments{
\item{x}{a \code{\link{gtypes}} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a data.frame with the columns 'id', 'strata', and two columns for every
  diploid locus, or one column of haplotypes.
}
\description{
Coerce a gtypes object to a data.frame.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{link{as.matrix.gtypes}}
}

