% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.gtypes.R
\docType{methods}
\name{as.matrix,gtypes-method}
\alias{as.matrix,gtypes-method}
\alias{as.matrix.gtypes}
\title{Convert \code{gtypes} To \code{matrix}}
\usage{
\S4method{as.matrix}{gtypes}(x, one.col = FALSE, sep = "/", ...)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}

\item{one.col}{logical. If \code{TRUE}, then result has one column per 
locus.}

\item{sep}{character to use to separate alleles if \code{one.col} is 
\code{TRUE}.}

\item{...}{additional arguments ot be passed to or from methods.}
}
\value{
A \code{matrix} or \code{data.frame} with one row per sample.
}
\description{
Convert genotypic data in the \code{@loci} slot of a 
  \linkS4class{gtypes} object to a \code{matrix}.
}
\author{
Eric Archer \email{eric.archer@noa.gov}
}
\seealso{
\link{df2gtypes}
}

