% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.matrix.gtypes}
\alias{as.matrix.gtypes}
\title{Coerce to a Matrix}
\usage{
\method{as.matrix}{gtypes}(x, ...)
}
\arguments{
\item{x}{a \code{\link{gtypes}} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A matrix with the columns 'id', 'strata', and two columns for every
  diploid locus, or one column of haplotypes
}
\description{
Coerce a gtypes object to a matrix.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{as.data.frame.gtypes}}
}

