% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.multidna.R
\name{as.multidna}
\alias{as.multidna}
\title{Convert to multidna}
\usage{
as.multidna(x)
}
\arguments{
\item{x}{a valid set of sequences: character matrix, list of 
character vectors, \code{\link{DNAbin}} object or list of them,
\linkS4class{gtypes} object, or \linkS4class{multidna} object.}
}
\description{
Convert a set of sequences to a multidna object if possible.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

