% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{bowhead.snp.position}
\alias{bowhead.snp.position}
\title{Bowhead Whale SNP Genotype Groups}
\format{data.frame}
\usage{
data(bowhead.snp.position)
}
\description{
A data.frame of position information for SNPs to be phased
}
\references{
Morin, P.A., Archer, F.I., Pease, V.L., Hancock-Hanser, 
  B.L., Robertson, K.M., Huebinger, R.M., Martien, K.K., Bickham, 
  J.W., George, J.C., Postma, L.D., Taylor, B.L., 2012. Empirical comparison 
  of single nucleotide polymorphisms and microsatellites for population and 
  demographic analyses of bowhead whales. 
  Endangered Species Research 19, 129-147.
}
\keyword{datasets}

