% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosability.R
\name{diagnosability}
\alias{diagnosability}
\title{Diagnosability}
\usage{
diagnosability(
  g,
  gene = 1,
  pairwise = FALSE,
  conf.level = 0.95,
  replace = FALSE,
  sampsize = NULL,
  train.pct = 0.5,
  min.n = 2,
  min.votes.pct = c(0.8, 0.9, 0.95),
  rp.nrep = 0,
  unk = NULL
)
}
\arguments{
\item{g}{haploid \code{\link{gtypes}} object with aligned sequences.}

\item{gene}{number or name of gene to use from multidna \code{@sequences} 
slot. Defaults to the first gene in the object.}

\item{pairwise}{do analysis on all pairwise combinations of strata?}

\item{conf.level}{confidence level for the \code{\link{binom.test}} 
confidence interval.}

\item{replace}{sample with replacement in Random Forest trees? 
(see \code{\link[randomForest]{randomForest}}).}

\item{sampsize}{sample size for each Random Forest tree? 
(see \code{\link[randomForest]{randomForest}}). If \code{NULL} a 
balanced sample size is chosen 
(see \code{\link[rfPermute]{balancedSampsize}}).}

\item{train.pct}{if \code{sampsize} is \code{NULL}, the percent of the 
minimum strata size to use for \code{sampsize}.}

\item{min.n}{minimum sample size across all strata.}

\item{min.votes.pct}{numeric vector giving the minimum percent of votes for
the assigned strata for a sample to be considered correctly assigned.}

\item{rp.nrep}{number of replicates for \code{rfPermute} computation of 
significance of site importance scores.}

\item{unk}{vector of strata to be treated as "unknowns" for prediction with 
Random Forest model.}
}
\value{
a list containing a data.frame of summary statistics (\code{smry}), 
  and the \code{randomForest} object (\code{rf}). If \code{pairwise} 
  is \code{TRUE} then the \code{rf} element is a list of 
  \code{randomForest} results for each row in \code{smry}.
}
\description{
Conduct Random Forest on a gtypes object to compute the
  diagnosability of each stratum (PD from Archer et al 2017).
}
\examples{
\dontrun{
library(strataG)
data(dloop.g)

pd <- diagnosability(dloop.g, pairwise = TRUE)

lapply(pd, function(x) x$rf.confusion.mat)
}

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
