% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosability.R
\name{diagnosability}
\alias{diagnosability}
\title{Diagnosability}
\usage{
diagnosability(
  g,
  gene = 1,
  pairwise = FALSE,
  conf.level = 0.95,
  unk = NULL,
  ...
)
}
\arguments{
\item{g}{haploid \code{\link{gtypes}} object with aligned sequences.}

\item{gene}{number or name of gene to use from multidna \code{@sequences} 
slot. Defaults to the first gene in the object.}

\item{pairwise}{do analysis on all pairwise combinations of strata?}

\item{conf.level}{confidence level for the \code{\link{binom.test}} 
confidence interval.}

\item{unk}{vector of strata to be treated as "unknowns" for prediction with 
Random Forest model.}

\item{...}{arguments passed to \code{\link[randomForest]{randomForest}}.}
}
\value{
a list containing a data.frame of summary statistics (\code{smry}), 
  and the \code{randomForest} object (\code{rf}). If \code{pairwise} 
  is \code{TRUE} then the \code{rf} element is a list of 
  \code{randomForest} results for each row in \code{smry}.
}
\description{
Conduct Random Forest on a gtypes object to compute the
  diagnosability of each stratum (PD from Archer et al 2017).
}
\examples{
\dontrun{
library(strataG)
data(dloop.g)

diagnosability(dloop.g, pairwise = TRUE)
}

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
