% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscRead.R
\name{fscRead}
\alias{fscRead}
\alias{fscReadArpFile}
\alias{fscExtractLoci}
\title{Read fastsimcoal output}
\usage{
fscRead(p, sim = 1)

fscReadArpFile(file)

fscExtractLoci(p, sim = 1, type = "all", gen.data = NULL,
  sep.chrom = FALSE, chrom = NULL)
}
\arguments{
\item{p}{list of fastsimcoal input parameters.}

\item{sim}{number of the simulation replicate to read.}

\item{file}{filename to read.}

\item{type}{type of marker to return.}

\item{gen.data}{matrix of parsed genetic data read from .arp file with 
`fscParseGeneticData()`.}

\item{sep.chrom}{return a list with chromosomes separated?}

\item{chrom}{numerical vector giving chromosomes to return. `NULL` 
returns all chromosomes.}
}
\description{
Read arlequin formatted output or parameter estimation files 
  generated by fastsimcoal
}
\note{
fastsimcoal is not included with `strataG` and must be downloaded 
  separately. Additionally, it must be installed such that it can be run from 
  the command line in the current working directory. See the vignette 
  for \code{external.programs} for installation instructions.
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time 
  coalescent simulator of genomic diversity under arbitrarily complex 
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  Excoffier, L., Dupanloup, I., Huerta-Sánchez, E., Sousa, V.C., 
  and M. Foll (2013) Robust demographic inference from genomic and SNP data. 
  PLOS Genetics, 9(10):e1003905. \cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}
\seealso{
\code{\link{fsc.input}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
