% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscWrite.R
\name{fscWrite}
\alias{fscWrite}
\alias{fscWriteSFS}
\title{Run fastsimcoal}
\usage{
fscWrite(demes, genetics, migration = NULL, events = NULL,
  label = "strataG.fastsimcoal")

fscWriteSFS(x, label)
}
\arguments{
\item{demes}{matrix of deme sampling information created by the 
\code{\link{fscSettingsDemes}} function.}

\item{genetics}{data.frame specifying loci to simulate created by the 
\code{\link{fscSettingsGenetics}} function.}

\item{migration}{a list of matrices giving the migration rates 
between pairs of demes created by the \code{\link{fscSettingsMigration}} 
function.}

\item{events}{matrix of historical events created by the 
\code{\link{fscSettingsEvents}} function.}

\item{label}{character string used to label output files for the simulation.}

\item{x}{a list of marginal and (optionally) joint site frequency spectra 
data frames. These can be the output of the \code{$sfs} element from 
\code{\link{fscReadSFSOutput}} or the output of \code{\link{sfs}} 
run with \code{fsc.dimnames = TRUE}.}
}
\value{
\describe{
 \item{fscWrite}{Writes input files for \code{fastsimcoal2} and returns a
   list of input parameters, input file, and input filenames. This list is 
   the primary input to \code{\link{fscRun}}.} 
 \item{fscWriteSFS}{Writes a set of \code{fastsimcoal2} formatted SFS files.}
 }
}
\description{
Run a fastsimcoal simulation and load results into a 
  \linkS4class{gtypes} object.
}
\note{
\code{fastsimcoal2} is not included with `strataG` and must be downloaded 
  separately. Additionally, it must be installed such that it can be run from 
  the command line in the current working directory. See the vignette 
  for \code{external.programs} for installation instructions.
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time 
  coalescent simulator of genomic diversity under arbitrarily complex 
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  Excoffier, L., Dupanloup, I., Huerta-Sánchez, E., Sousa, V.C., 
  and M. Foll (2013) Robust demographic inference from genomic and SNP data. 
  PLOS Genetics, 9(10):e1003905. \cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}
\seealso{
\code{\link{fsc.input}}, \code{\link{fscRun}}, 
 \code{\link{fscRead}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
