% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fst.to.nm}
\alias{fst.to.nm}
\title{Nm from Fst}
\usage{
fst.to.nm(fst, diploid = TRUE)
}
\arguments{
\item{fst}{estimate of Fst between populations.}

\item{diploid}{logical. Is Fst from diploid (TRUE) or haploid (FALSE) data?}
}
\description{
Calculate Nm for a given value of Fst.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

