% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusFs.R
\name{fusFs}
\alias{fusFs}
\title{Fu's Fs}
\usage{
fusFs(x)
}
\arguments{
\item{x}{set of DNA sequences or a haploid \linkS4class{gtypes} 
object with sequences.}
}
\description{
Calculate Fu's Fs for a set of sequences to test 
  for selection.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Fu, Y-X. 1997. Statistical tests of neutrality of mutations 
  against population growth, hitchiking and background selection. 
  Genetics 147:915-925.
}

