% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypes2loci.R
\name{gtypes2loci}
\alias{gtypes2loci}
\title{Convert Between \code{gtypes} And \code{loci} objects.}
\usage{
gtypes2loci(x)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}
}
\description{
Convert a \code{gtypes} object to a \code{\link[pegas]{loci}} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\link{initialize.gtypes}, \link{df2gtypes}, \link{sequence2gtypes}, 
  \link{gtypes2df}, \link{gtypes2genind}
}

