% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.diploid.R, R/is.haploid.R
\name{is.diploid}
\alias{is.diploid}
\alias{is.haploid}
\title{Check Ploidy}
\usage{
is.diploid(g)

is.haploid(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Check if a gtypes object contains haploid or diploid data.
}
\details{
Checks if the \code{locus.data} element of \code{g} is one column (haploid)
  or an even-numbered column (diploid) matrix.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

