% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldNe.R
\name{ldNe}
\alias{ldNe}
\title{ldNe}
\usage{
ldNe(g, maf.threshold = 0)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{maf.threshold}{}
}
\value{
Ne estimate
}
\description{
Estimate Ne from linkage disequilibrium
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Waples
}

