% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldNe.R
\name{ldNe}
\alias{ldNe}
\title{ldNe}
\usage{
ldNe(g, maf.threshold = 0)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{maf.threshold}{smallest minimum allele frequency permitted to include 
a locus in calculation.}
}
\value{
vector with sample size (\code{S}), number of pairwise loci comparisons 
  used (\code{num.comp}), mean r-squared (\code{mean.rsq}), and 
  Ne estimate (\code{Ne}).
}
\description{
Estimate Ne from linkage disequilibrium based on Pearson correlation 
  approximation as in Waples et al 2016.
}
\author{
Eric Archer \email{eric.archer@noaa.gov} (adapted from code by R. Waples)
}
\references{
Waples, R.S. 2006. A bias correction for estimates of effective population 
  size based on linkage disequilibrium at unlinked gene loci. 
  Conservation Genetics 7:167-184. \cr
  Waples RK, Larson WA, and Waples RS. 2016. 
  Estimating contemporary effective population size in non-model species using 
  linkage disequilibrium across thousands of loci. Heredity 117:233-240; 
  doi:10.1038/hdy.2016.60
}

