% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega.R
\name{mega}
\alias{MEGA}
\alias{mega}
\alias{mega,}
\alias{read.mega}
\alias{write.mega}
\title{Read and Write MEGA}
\usage{
read.mega(file)

write.mega(g, file = NULL, title = NULL, line.width = 60, locus = 1)
}
\arguments{
\item{file}{a MEGA-formatted file of sequences.}

\item{g}{a \linkS4class{gtypes} object.}

\item{title}{title for data in file.}

\item{line.width}{width of sequence lines.}

\item{locus}{number or name of locus to write.}
}
\value{
for \code{read.mega}, a list of:
\tabular{ll}{
  \code{title} \tab title of MEGA file.\cr
  \code{dna.seq} \tab a list of DNA sequences.\cr
}
}
\description{
Read and write MEGA formatted files.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

