% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nucleotide.diversity.R
\name{nucleotide.diversity}
\alias{nucleotide.diversity}
\title{Nucleotide Diversity}
\usage{
nucleotide.diversity(x, bases = c("a", "c", "g", "t"))
}
\arguments{
\item{x}{a list of sequences or a \code{\link{gtypes}} object.}

\item{bases}{nucleotides to consider when calculating diversity.}
}
\value{
Nucleotide diversity by site.
}
\description{
Calculate nucleotide diversity for set of haplotypes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

