% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/num.alleles.R
\name{num.alleles}
\alias{num.alleles}
\title{Number of Alleles}
\usage{
num.alleles(g, use.na.rows = TRUE)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{use.na.rows}{logical. FALSE deletes any row with an NA for either allele.
  TRUE uses all unique (non-NA) alleles in frequency.}
}
\value{
vector of number of alleles per locus.
}
\description{
Return the number of alleles for each locus.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

