% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num.alleles.shared}
\alias{num.alleles.shared}
\title{Number of Shared Alleles}
\usage{
num.alleles.shared(id1, id2, g)
}
\arguments{
\item{id1,id2}{ids of samples to compare.}

\item{g}{a \code{\link{gtypes}} object.}
}
\description{
Calculate the number of alleles shared between two samples at each locus.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

