% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numAlleles.R
\name{numAlleles}
\alias{numAlleles}
\title{Number of Alleles}
\usage{
numAlleles(g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}
}
\value{
vector of number of alleles per locus.
}
\description{
Return the number of alleles for each locus.
}
\examples{
data(msats.g)

numAlleles(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

