% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericSNPmat.R
\name{numericSNPmat}
\alias{numericSNPmat}
\title{Create a numeric SNP matrix}
\usage{
numericSNPmat(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a data.frame with all biallelic loci recoded.
}
\description{
Create a matrix of SNPs coded as 0, 1, 2, where 0 and 2 are 
  the two homozygotes and 1 is the heterozygote.
}
\note{
\code{g} must be a diploid \code{gtypes} object and have at least one 
  locus with one or two alleles.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
