% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numGenotyped.R
\name{numGenotyped}
\alias{numGenotyped}
\title{Number of Individuals Genotyped}
\usage{
numGenotyped(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
vector of number of alleles per locus.
}
\description{
Return the number of individuals genotyped for each locus.
}
\examples{
data(msats.g)

numGenotyped(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
