% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numGensEq.R
\name{numGensEq}
\alias{numGensEq}
\title{Number of Generations to Equilibrium}
\usage{
numGensEq(fst, Ne, gen.time)
}
\arguments{
\item{fst}{estimate of Fst between populations.}

\item{Ne}{population size.}

\item{gen.time}{generation time.}
}
\description{
Calculate the number of generations to equilibrium based on a an
  ideal Wright model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

