% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numMissing.R
\name{numMissing}
\alias{numMissing}
\title{Number Missing Data}
\usage{
numMissing(g, prop = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{prop}{logical determining whether to return proportion missing.}
}
\value{
a vector of loci with number (or, if \code{prop = TRUE},
  the proportion) of individuals missing data for at least one allele.
}
\description{
Calculate the number of individuals with missing data by locus.
}
\examples{
data(msats.g)

numMissing(msats.g)
numMissing(msats.g, prop = TRUE)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
