% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pairwise.shared.loci.R
\name{pairwise.shared.loci}
\alias{pairwise.shared.loci}
\title{Calculate Pairwise Shared Loci}
\usage{
pairwise.shared.loci(g, num.cores = 1)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{num.cores}{number of CPU cores to use. Value is passed to \code{\link[parallel]{mclapply}}.}
}
\value{
matrix summary of pairwise shared loci.
}
\description{
Calculate number of loci shared between pairs of individuals.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{num.alleles.shared}}
}

