% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popStructTest.R
\name{popStructTest}
\alias{overallTest}
\alias{pairwiseTest}
\alias{popStructTest}
\alias{statList}
\title{Population Differentiation Tests}
\usage{
popStructTest(g, nrep = 1000, stats = "all", type = c("both", "overall",
  "pairwise"), keep.null = FALSE, quietly = FALSE, write.output = FALSE,
  ...)

overallTest(g, nrep = 1000, stats = "all", keep.null = FALSE,
  quietly = FALSE, ...)

pairwiseTest(g, nrep = 1000, stats = "all", keep.null = FALSE,
  quietly = FALSE, ...)

statList(stats = "all")
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{nrep}{number specifying number of permutation replicates to use for 
permutation test.}

\item{stats}{a character vector or list of functions specifying which 
anlayses to conduct. If characters, then valid possible choices are: 
"phist", "fst", "fst.prime", "fis", "gst", "gst.prime", "gst.dbl.prime", "d", 
or "chi2", or "all". If a list, then functions must be a valid population 
structure function (see \code{\link{popStructStat}}) taking a 
\linkS4class{gtypes} object and returning a named statistic estimate.}

\item{type}{character determining type of test to conduct. Can be "overall", 
"pairwise", or "both". If "pairwise" or "both" are chosen and there are 
only two strata, then only an overall test will be conducted.}

\item{keep.null}{logical. Keep the null distribution from the 
permutation test?}

\item{quietly}{logical. Print progress and results?}

\item{write.output}{logical. Write a .csv file with results?}

\item{...}{other parameters to be passed to population 
differentiation functions.}
}
\value{
\describe{
 \item{overall}{a list containing:
   \tabular{ll}{
     \code{strata.freq} \tab a vector of the sample sizes for each stratum.\cr
     \code{result} \tab a matrix with the statistic estimate and p-value 
       for each statistic.\cr
     \code{null.dist} \tab a matrix with the null distributions for 
       each statistic.\cr
   }}
 \item{pairwise}{a list containing:
   \tabular{ll}{
     \code{result} \tab a data.frame with the result of each pairwise 
       comparison on a line.\cr
     \code{pair.mat} \tab a list with a pairwise matrix for each statistic. 
       Values in lower left are the statistic estimate, and upper right are p-values.\cr
     \code{null.dist} \tab a matrix with the null distributions for 
       each statistic.\cr
   }}
}
}
\description{
Conduct overall and/or pairwise tests of 
  population differentiation.
}
\examples{
data(msats.g)
msats.g <- stratify(msats.g, "fine")

# Conduct an overall Chi-squared test
ovl <- overallTest(msats.g, stats = "chi2")

# Conduct a pairwise test for Gst
pws <- pairwiseTest(msats.g, stats = list(statGst))

# Conduct both overall and pairwise tests for Fst and F'st
full <- popStructTest(msats.g, stats = c("fst", "fst.prime"))
print(full$overall)
print(full$pairwise)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

