% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.gtypeSummary.R
\name{print.gtypeSummary}
\alias{print.gtypeSummary}
\title{Print gtypeSummary}
\usage{
\method{print}{gtypeSummary}(x, ... )
}
\arguments{
\item{x}{list from summary.gtypes}

\item{...}{other arguments (ignored)}
}
\description{
Print the output for a \code{gtypeSummary} object from \code{\link{summary.gtypes}}.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

