% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propUniqueAlleles.R
\name{propUniqueAlleles}
\alias{propUniqueAlleles}
\title{Proportion Unique Alleles}
\usage{
propUniqueAlleles(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a vector of the proportion of unique (occuring only in one individual) 
  alleles for each locus.
}
\description{
Calculate the proportion of alleles that are unique.
}
\note{
If \code{g} is a haploid object with sequences, make sure to run 
  \code{\link{labelHaplotypes}} if sequences aren't already grouped by 
  haplotype.
}
\examples{
data(msats.g)

propUniqueAlleles(msats.g)

}
\seealso{
\link{privateAlleles}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
