% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGenData.R
\name{readGenData}
\alias{readGenData}
\title{Read Genetic Data}
\usage{
readGenData(file, na.strings = c(NA, "NA", "", " ", "?", "."), ...)
}
\arguments{
\item{file}{filename of .csv file.}

\item{na.strings}{see \code{\link[data.table]{fread}}.}

\item{...}{other arguments passed to \code{\link[data.table]{fread}}.}
}
\value{
a \code{data.frame}.
}
\description{
A wrapper for \code{\link[data.table]{fread}} that sets 
  common values for missing data and removes blank lines.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
