% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restratify.R
\name{restratify}
\alias{restratify}
\title{Restratify gtypes}
\usage{
restratify(g, new.strata)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{new.strata}{a character or numeric vector specifying new strata for samples.
  If vector is named, names are assumed to be sample IDs. If not named, vector must be same length as number
  of samples in \code{g}.}
}
\value{
new \code{\link{gtypes}} object.
}
\description{
Change the stratification of samples in a \code{\link{gtypes}} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

