% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequence.summary.R
\name{sequence.summary}
\alias{sequence.summary}
\title{Sequence Summary}
\usage{
sequence.summary(x)
}
\arguments{
\item{x}{a haploid \code{\link{gtypes}} object with sequences or a list of DNA sequences.}
}
\description{
Summaries for each sequence.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

