% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{sfs}
\alias{sfs}
\title{Site Frequency Spectrum}
\usage{
sfs(g, ref.allele = NULL, joint = FALSE, strata = NULL,
  simplify = TRUE, na.action = c("fail", "filter", "resample"))
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{ref.allele}{an optional vector of reference alleles for each SNP. 
See \code{\link{as.data.frame.gtypes}} for more information.}

\item{joint}{a logical specifying whether or not to compute joint (2D) 
SFS.}

\item{strata}{a vector of strata to use to compute joint SFS. 
If `NULL` (default), then all pairs of strata in `g` are used.}

\item{simplify}{a logical specifying whether to return a matrix if only 
a single joint SFS is computed (one pair of strata).}

\item{na.action}{action to take if genotypes are missing for some samples. 
If \code{"fail"}, an error is thrown if any genotypes are missing. If 
\code{"filter"}, SNPs with missing genotypes are removed. If 
\code{"resample"}, the minimum number of genotyped individuals across 
all SNPs is first determined, then this number of non-missing genotypes 
are randomly selected for each SNP to calculate the SFS.}
}
\value{
Either a named vector of the 1D site frequency 
  spectrum if \code{joint = FALSE}, or a list of matrices of the 2D
  site frequency spectra for pairs of strata if \code{joint = TRUE}.
}
\description{
Calculate the SFS from a \linkS4class{gtypes} object for 
  all individuals or pairs of strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
