% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stat.gst.dbl.prime}
\alias{stat.gst.dbl.prime}
\title{G''st Analysis of Population Structure}
\usage{
stat.gst.dbl.prime(g)
}
\arguments{
\item{g}{a \code{gtypes} object}
}
\description{
G''st Analysis of Population Structure
}
\details{
Calculate G''st from Meirmans and Hedrick 2010
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

stat.gst.dbl.prime(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

