% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat.gst.prime.nei.R
\name{stat.gst.prime.nei}
\alias{stat.gst.prime.nei}
\title{Nei's G'st Analysis of Population Structure}
\usage{
stat.gst.prime.nei(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\description{
Nei's G'st Analysis of Population Structure
}
\examples{
data(dolph.strata)
data(dolph.msats)
msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

stat.gst.prime.nei(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Nei 1987
}

