% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stopifnot.gtype.struct.func.R
\name{stopifnot.gtype.struct.func}
\alias{stopifnot.gtype.struct.func}
\title{Check gtype Population Structure Function}
\usage{
stopifnot.gtype.struct.func(func)
}
\arguments{
\item{func}{function to test.}
}
\description{
Stop execution if function is not a valid population structure
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

