% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{package}
\name{strataG-package}
\alias{strataG}
\alias{strataG-package}
\title{Summaries and population structure analyses of haplotypic and genotypic data}
\description{
strataG
}
\keyword{package}

