% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{package}
\name{strataG-package}
\alias{strataG-package}
\alias{strataG}
\title{Summaries and population structure analyses of DNA sequence genotypic 
  data}
\description{
strataG
}
\keyword{package}
