% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gtypes.R
\docType{methods}
\name{summary,gtypes-method}
\alias{print.gtypeSummary}
\alias{summary}
\alias{summary,gtypes-method}
\title{Summarize gtypes Object}
\usage{
\S4method{summary}{gtypes}(object, ...)

\method{print}{gtypeSummary}(x, ...)
}
\arguments{
\item{object}{a \linkS4class{gtypes} object.}

\item{...}{other arguments (ignored).}

\item{x}{list from summary.gtypes}
}
\value{
a list with the following elements:
\describe{
  \item{\code{num.ind}}{number of individuals}
  \item{\code{num.loc}}{number of loci}
  \item{\code{num.strata}}{number of strata}
  \item{\code{unstratified}}{number of unstratified samples}
  \item{\code{schemes}}{names of stratification schemes}
  \item{\code{allele.freqs}}{a list with tables of allele frequencies by strata}
  \item{\code{strata.smry}}{a by-strata data.frame summarizing haplotypes or loci}
  \item{\code{locus.smry}}{a data.frame summarizing each locus for 
    non-haploid objects, \code{NULL} for haploid objects}
  \item{\code{seq.smry}}{a summary of the sequence length and base frequencies}
}
}
\description{
Generate a summary of a \code{gtypes} object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

