% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{Theta}
\usage{
theta(g, by.strata = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{logical - return results grouped by strata?}
}
\value{
vector of theta values for each locus.
}
\description{
Calculate theta from heterozygosity of each locus.
}
\details{
Calculates theta for each locus using the
  \code{\link[pegas]{theta.h}} function.
}
\examples{
data(msats.g)

theta(msats.g)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
