% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimNs.R
\name{trimNs}
\alias{trimNs}
\title{Trim N's From Sequences}
\usage{
trimNs(x)
}
\arguments{
\item{x}{a \code{\link[ape]{DNAbin}} object.}
}
\value{
sequences with beginning and trailing N's removed.
}
\description{
Removes N's from beginning and end of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

