% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.snapp.nexus}
\alias{write.snapp.nexus}
\title{Write NEXUS File for SNAPP}
\usage{
write.snapp.nexus(g, file = "snapp.data.nex")
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{file}{the filename the NEXUS file to output.}
}
\description{
Write NEXUS File for SNAPP
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

